[feat] Migrate all status related endpoint
