+++ templates/new_status.html
@@ -1,7 +1,7 @@
-<div class="status_bar" hx-trigger="done" hx-get='/complete-update' hx-swap='outerHTML' hx-target='this'>
+<div class="status_bar" hx-trigger="done" hx-get='/books/complete' hx-swap='outerHTML' hx-target='this'>
     <h3 role="status" id="pblabel" tabindex="-1" autofocus>{{ status }}</h3>
 
-    <div hx-get="/book_status/{{all_unique_books|length}}" hx-trigger="every 1s" hx-target='.progress'>
+    <div hx-get="/books/status/{{all_unique_books|length}}" hx-trigger="every 1s" hx-target='.progress'>
         <div class='flex mx-auto justify-center items-center text-blue-700 w-3/4'>
             <div class="progress" role="progressbar" aria-valuenow={{ progress }} style=width:{{ progress }}%>
             </div>
