+++ templates/partials/nb_lib_btn.html
@@ -5,7 +5,7 @@
         </summary>
         <ul class="bg-sky-700 text-sm text-white w-56">
             <li>
-                <a class="py-0.5 hover:text-white hover:bg-sky-500" hx-get="/update_header" hx-swap="outerHTML"
+                <a class="py-0.5 hover:text-white hover:bg-sky-500" hx-get="/nav" hx-swap="outerHTML"
                     hx-target="#header" hx-indicator="#head-spinner">
                     Update Book Count
                 </a>
