+++ templates/partials/search_table.html
@@ -7,7 +7,7 @@
     {% set nav_search = '/search/results/?' %}
     {% endif %}
 
-    <form id="myForm" hx-post="/ingest_books_navbar" hx-indicator="#head-spinner" hx-target="#header"
+    <form id="myForm" hx-post="/books" hx-indicator="#head-spinner" hx-target="#header"
         hx-swap="outerHTML">
 
         <div class="join flex items-center justify-left">
