+++ templates/update_status.html
@@ -1,4 +1,4 @@
-<div class="status_bar" hx-trigger="done" hx-get='complete-update' hx-swap='outerHTML' hx-target='.this'>
+<div class="status_bar" hx-trigger="done" hx-get='/books/complete' hx-swap='outerHTML' hx-target='.this'>
     <div hx-get="/book_status/{{all_unique_books|length}}" hx-trigger="every 1s" hx-target='.book_text'>
         <div class='flex mx-auto justify-center items-center text-blue-700 w-3/4'>
             <div class="book_text"></div>
